requirement.txt
spacy
pandas
torch
torchvision
numpy
Pillow
torchfile
sklearn
vit-pytorch
